requirements.txt
altair==4.1.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
fuzzywuzzy==0.18.0
gitdb==4.0.9
GitPython==3.1.18
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.2
Jinja2==3.0.3
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
lxml==4.8.0
MarkupSafe==2.0.1
mistune==0.8.4
mpmath==1.2.1
mysql-connector==2.2.9
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.19.5
openpyxl==3.0.9
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.3.1
plotly==5.6.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
ptyprocess==0.7.0
pyarrow==6.0.1
pycparser==2.21
pydeck==0.6.2
Pygments==2.11.2
Pympler==1.0.1
PyMySQL==1.0.2
pyparsing==3.0.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
requests==2.27.1
scipy==1.5.4
semver==2.13.0
Send2Trash==1.8.0
setuptools==59.6.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
statsmodels==0.12.2
stqdm==0.0.4
streamlit==1.8.1
streamlit-option-menu==0.3.2
sympy==1.9
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wikipedia==1.4.0
xlrd==2.0.1
zipp==3.6.0
